MFC standardization of the ELF types and update of the ELF constants: 	lib/libc/gen/nlist.c:1.19 	libexec/rtld-elf/rtld.c:1.108 	libexec/rtld-elf/rtld.h:1.36 	libexec/rtld-elf/alpha/reloc.c:1.21 	libexec/rtld-elf/amd64/reloc.c:1.16 	libexec/rtld-elf/ia64/reloc.c:1.16 	libexec/rtld-elf/sparc64/reloc.c:1.11 	share/man/man5/elf.5:1.32-1.33 	sys/alpha/alpha/elf_machdep.c:1.20 	sys/amd64/amd64/elf_machdep.c:1.24 	sys/boot/common/bootstrap.h:1.42 	sys/boot/common/load_elf.c:1.33 	sys/boot/common/load_elf_obj.c:1.2 	sys/boot/common/reloc_elf.c:1.2 	sys/ia64/ia64/elf_machdep.c:1.21 	sys/kern/link_elf.c:1.87 	sys/kern/link_elf_obj.c:1.90 	sys/sparc64/sparc64/elf_machdep.c:1.20 	sys/sys/elf32.h:1.9-1.10 	sys/sys/elf64.h:1.11-1.13 	sys/sys/elf_common.h:1.16 	sys/sys/elf_generic.h:1.7 	sys/sys/imgact_elf.h:1.28 	sys/sys/linker.h:1.40 	usr.bin/elf2aout/elf2aout.c:1.10 	usr.bin/elfdump/elfdump.c:1.13 	usr.sbin/crunch/crunchide/exec_elf32.c:1.15 	usr.sbin/kldxref/ef.c:1.9 	usr.sbin/kldxref/ef.h:1.6 	usr.sbin/kldxref/ef_amd64.c:1.3 	usr.sbin/kldxref/ef_i386.c:1.3 	usr.sbin/kldxref/ef_obj.c:1.4 	usr.sbin/kldxref/ef_powerpc.c:1.3 	usr.sbin/kldxref/ef_sparc64.c:1.4-1.5

+++ libexec/rtld-elf/amd64/reloc.c
@@ -312,9 +312,9 @@ reloc_non_plt(Obj_Entry *obj, Obj_Entry *obj_rtld)
 	/* missing: R_X86_64_GOTPCREL, R_X86_64_32, R_X86_64_32S, R_X86_64_16, R_X86_64_PC16, R_X86_64_8, R_X86_64_PC8 */
 
 	    default:
-		_rtld_error("%s: Unsupported relocation type %d"
+		_rtld_error("%s: Unsupported relocation type %u"
 		  " in non-PLT relocations\n", obj->path,
-		  ELF_R_TYPE(rela->r_info));
+		  (unsigned int)ELF_R_TYPE(rela->r_info));
 		goto done;
 	    }
 	}
